+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed:
+- Navbar changed to Salomon style
+
 ### Fixed:
 - Navigation glitch where the bottom nav bar was moving
 - Error handling when page does not exist